Add LAION Aesthetic metric v1
Add LAION Aesthetic metric which is based on the
https://github.com/LAION-AI/aesthetic-predictor. New metric takes an
image as a torch.Tensor as an input, processes it with a given CLIP
model, and finally applies a pretrained linear head to compute an
aesthetics score. Aesthetic score ranges between 1.0 and 10.0.

Add tests for all 3 supported models: vit_l_14, vit_b_16, vit_b_32.