feat: aesthetic laion metric (#424)
* Add LAION Aesthetic metric v1

Add LAION Aesthetic metric which is based on the
https://github.com/LAION-AI/aesthetic-predictor. New metric takes an
image as a torch.Tensor as an input, processes it with a given CLIP
model, and finally applies a pretrained linear head to compute an
aesthetics score. Aesthetic score ranges between 1.0 and 10.0.

Add tests for all 3 supported models: vit_l_14, vit_b_16, vit_b_32.

* Update docs

* Fix torch.device issue

Aesthetic model was loaded on the CPU and never moved to the selected
device

* Add a test case with device="cuda"

* Remove CLIPVariantAesthetics enum

Use a dict to map hugging face model name to a aesthetic linear head
name instead of the enum.
Update related test cases.
Add extra test for parameter validation.

* Fix TypeError: clip_model_name -> model_name_or_path

* Move aesthetic model predictions to cpu

* Change _get_aesthetic_model default param

* Add map_location to torch.load

* Fix linting errors

Docs for the __init__ method did contain several errors

* Fix typo in the filename: liaon -> laion

* refactor: small fixes

---------

Co-authored-by: Kirill Ivanov <kirill.dm.ivanov@gmail.com>